+++ src/components/AdminSettings.vue
@@ -30,12 +30,12 @@
 				:has-error="!hasEnabledSupportedUserOidcApp || hasOidcAppErrorWithNextcloudHub" />
 			<ErrorNote
 				v-if="!hasEnabledSupportedUserOidcApp"
-				:error-title="messagesFmt.appNotEnabledOrUnsupported('user_oidc', getMinSupportedUserOidcVersion)"
+				:error-title="messagesFmt.appNotEnabledOrUnsupported(getUserOidcAppName, getMinSupportedUserOidcVersion)"
 				:error-link="appLinks.user_oidc.installLink"
 				:error-link-label="messages.installLatestVersionNow" />
 			<ErrorNote
 				v-else-if="hasOidcAppErrorWithNextcloudHub"
-				:error-title="messagesFmt.appNotEnabledOrUnsupported('oidc', getMinSupportedOidcVersion)"
+				:error-title="messagesFmt.appNotEnabledOrUnsupported(getOidcAppName, getMinSupportedOidcVersion)"
 				:error-link="appLinks.oidc.installLink"
 				:error-link-label="messages.installLatestVersionNow" />
 			<div class="authorization-settings--content">
@@ -58,7 +58,7 @@
 					<div class="error-container">
 						<ErrorLabel
 							v-if="!hasEnabledSupportedOIDCApp"
-							:error="messagesFmt.appNotEnabledOrUnsupported('oidc', getMinSupportedOidcVersion)"
+							:error="messagesFmt.appNotEnabledOrUnsupported(getOidcAppName, getMinSupportedOidcVersion)"
 							:disabled="disableNCHubUnsupportedHint" />
 					</div>
 					<NcCheckboxRadioSwitch
@@ -289,7 +289,7 @@
 				:is-dark-theme="isDarkTheme" />
 			<ErrorNote
 				v-if="showGroupfoldersAppError"
-				:error-title="messagesFmt.appNotEnabledOrUnsupported('groupfolders', getMinSupportedGroupfoldersVersion)"
+				:error-title="messagesFmt.appNotEnabledOrUnsupported(getGroupfoldersAppName, getMinSupportedGroupfoldersVersion)"
 				:error-link="appLinks.groupfolders.installLink"
 				:error-link-label="messages.installLatestVersionNow" />
 			<NcNoteCard v-else-if="projectFolderSetupError || isThereErrorAfterProjectFolderAndAppPasswordSetup" class="note-card" type="error">
@@ -698,12 +698,12 @@ export default {
 			return t('integration_openproject', 'Setting up the OpenProject user, group and team folder was not possible. Please check this {htmlLink} on how to resolve this situation.', { htmlLink }, null, { escape: false, sanitize: false })
 		},
 		getAdminAuditConfigurationHint() {
-			const linkTextForAdminAudit = t('integration_openproject', 'Admin Audit')
+			const linkTextForAdminAudit = t('integration_openproject', this.getAdminAuditAppName)
 			const adminAuditAppUrlForDownload = generateUrl('settings/apps/featured/admin_audit')
 			const linkTextForDocumentation = t('integration_openproject', 'documentation')
 			const htmlLinkForAdminAudit = `<a class="link" href="${adminAuditAppUrlForDownload}" target="_blank" title="${linkTextForAdminAudit}">${linkTextForAdminAudit}</a>`
 			const htmlLinkForDocumentaion = `<a class="link" href="https://docs.nextcloud.com/server/latest/admin_manual/configuration_server/logging_configuration.html#admin-audit-log-optional" target="_blank" title="${linkTextForDocumentation}">${linkTextForDocumentation}</a>`
-			const hintTextForAdminAudit = t('integration_openproject', 'To activate audit logs for the OpenProject integration, please enable the {htmlLinkForAdminAudit} app and follow the configuration steps outlined in the {htmlLinkForDocumentaion}.', { htmlLinkForAdminAudit, htmlLinkForDocumentaion }, null, { escape: false, sanitize: false })
+			const hintTextForAdminAudit = t('integration_openproject', 'To activate audit logs for the OpenProject integration, please enable the "{htmlLinkForAdminAudit}" app and follow the configuration steps outlined in the {htmlLinkForDocumentaion}.', { htmlLinkForAdminAudit, htmlLinkForDocumentaion }, null, { escape: false, sanitize: false })
 			return dompurify.sanitize(hintTextForAdminAudit, { ADD_ATTR: ['target'] })
 		},
 		getGroupFoldersEncryptionWarningHint() {
@@ -785,6 +785,18 @@ export default {
 		},
 		getSSOProviderType() {
 			return this.authorizationSetting.SSOProviderType
+		},
+		getUserOidcAppName() {
+			return this.state.apps.user_oidc.name
+		},
+		getOidcAppName() {
+			return this.state.apps.oidc.name
+		},
+		getGroupfoldersAppName() {
+			return this.state.apps.groupfolders.name
+		},
+		getAdminAuditAppName() {
+			return this.state.admin_audit_app_name
 		},
 		hasEnabledSupportedUserOidcApp() {
 			return this.state.apps.user_oidc.enabled && this.state.apps.user_oidc.supported